- Remove pure Python BigInteger implementation - update code for improved C-extension bindings - update documentation for BigInteger (relocate) - fix a couple of minor documentation issues - fix some linting
