+++ docs/source/library/core/bloomfilter.rst
@@ -3,7 +3,7 @@
 
 While ready to install 3rd party bloom filter implementations exist on PyPi, they unfortunately do not comply with the NEO implementation.
 
-Bloomfilter's can be broadcasted over the network and thus there is a need to have identifical results.
+Bloomfilter's can be broadcasted over the network and thus there is a need to have identical results.
 
 
 .. automodule:: neo3.core.cryptography.bloomfilter
