+++ docs/source/library/index.rst
@@ -14,3 +14,4 @@
    core/index
    network/index
    storage/index
+   vm/index
