+++ docs/source/library/vm/biginteger.rst
@@ -1,7 +1,9 @@
+.. _library-vm-biginteger:
+
 :mod:`biginteger` --- A C# like BigInteger implementation
 =========================================================
 
-While Python natively supports infinite numbers through its build in `int <https://docs.python.org/3/library/functions.html#int>`_ type, the BigInteger class of the reference implementation exhibits different behaviour in certain areas. By using this class we can prevent any discrepancies in price calculations, virtual machine execution path and more.
+While Python natively supports infinite numbers through its build in `int <https://docs.python.org/3/library/functions.html#int>`_ type, the BigInteger class of the C# reference implementation exhibits different behaviour in certain areas. By using this class we can prevent any discrepancies in price calculations, virtual machine execution path and more.
 
 In particular the following differences in behaviour can be expected when using this class.
 
@@ -11,11 +13,11 @@ In particular the following differences in behaviour can be expected when using 
   - A negative right shift becomes a positive left shift.
 
 - Increased modulo rounding precision.
+- Use modulo truncated division instead of floored division.
 - Serialization of negative numbers can in certain cases return an extra byte.
 - Integer division instead of floating point division.
 
-.. automodule:: neo3.core.types.biginteger
+.. autoclass:: neo3.vm.BigInteger
    :members:
    :undoc-members:
    :show-inheritance:
-
