+++ docs/source/library/vm/index.rst
@@ -1,0 +1,12 @@
+.. _library-vm-index:
+
+***************
+Virtual Machine
+***************
+
+The VM package holds the modules related to the NEO Virtual Machine.
+
+.. toctree::
+   :maxdepth: 2
+
+   biginteger
