+++ neo3/contracts/native/__init__.py
@@ -2,4 +2,4 @@ from .nativecontract import (NativeContract, CallFlags)
 
 __all__ = ['NativeContract',
            'CallFlags'
-]
+           ]
