+++ neo3/core/types/__init__.py
@@ -1,4 +1,3 @@
-from .biginteger import BigInteger
 from .uint import *
 
-__all__ = ['BigInteger', 'UInt160', 'UInt256']
+__all__ = ['UInt160', 'UInt256']
